Rolloff: In case of a tie, pick winner randomly
All modern JavaScript implementations use stable sorts. This means
that relative ordering of equal items is maintained.

The previous implementation of the bidding was sorting the bids and
then picking the first element. In cases of ties, this results in the
first item in the array with that amount always being picked. The correct solution
is to pick randomly between the players with the same bid.